Move the Click-Cooper tests into their own module
They were getting really big, so having them in a separate file is just
good for our sanity.
